+++ phonopy/harmonic/dynamical_matrix.py
@@ -470,7 +470,11 @@ class DynamicalMatrixNAC(DynamicalMatrix):
     @property
     def nac_params(self):
         """Return NAC basic parameters."""
-        return {"born": self.born, "factor": self.factor, "dielectric": self.dielectric}
+        return {
+            "born": self.born,
+            "factor": self.nac_factor,
+            "dielectric": self.dielectric_constant,
+        }
 
     @nac_params.setter
     def nac_params(self, nac_params):
@@ -543,6 +547,7 @@ class DynamicalMatrixGL(DynamicalMatrixNAC):
         force_constants,
         nac_params=None,
         num_G_points=None,  # For Gonze NAC
+        with_full_terms=False,
         decimals=None,
         symprec=1e-5,
         log_level=0,
@@ -561,6 +566,9 @@ class DynamicalMatrixGL(DynamicalMatrixNAC):
             shape=(supercell atoms, supercell atoms, 3, 3) for full FC.
             shape=(primitive atoms, supercell atoms, 3, 3) for compact FC.
             dtype='double'
+        with_full_terms : bool, optional
+            When False, only reciprocal terms are considered for NAC. This is the
+            default and the reasonable choice.
         symprec : float, optional, defualt=1e-5
             Symmetri tolerance.
         decimals : int, optional, default=None
@@ -580,6 +588,7 @@ class DynamicalMatrixGL(DynamicalMatrixNAC):
 
         # For the method by Gonze et al.
         self._Gonze_force_constants = None
+        self._with_full_terms = with_full_terms
         if num_G_points is None:
             self._num_G_points = 300
         else:
@@ -588,6 +597,9 @@ class DynamicalMatrixGL(DynamicalMatrixNAC):
         self._G_cutoff = None
         self._Lambda = None  # 4*Lambda**2 is stored.
         self._dd_q0 = None
+        self._dd_real_q0 = None
+        self._dd_limiting = None
+        self._H = None
 
         if nac_params is not None:
             self.nac_params = nac_params
@@ -633,8 +645,6 @@ class DynamicalMatrixGL(DynamicalMatrixNAC):
             GeG = self._G_cutoff**2 * np.trace(self._dielectric) / 3
             self._Lambda = np.sqrt(-GeG / 4 / np.log(exp_cutoff))
 
-        # self._H = self._get_H()
-
     def make_Gonze_nac_dataset(self):
         """Prepare Gonze-Lee force constants.
 
@@ -652,6 +662,11 @@ class DynamicalMatrixGL(DynamicalMatrixNAC):
                 "implemented."
             )
             sys.exit(1)
+
+        if self._with_full_terms:
+            self._run_limiting_dipole_dipole()
+            self._H = self._get_H()
+            self._run_real_dipole_dipole_q0()
 
         fc_shape = self._force_constants.shape
         d2f = DynmatToForceConstants(
@@ -707,6 +722,7 @@ class DynamicalMatrixGL(DynamicalMatrixNAC):
         self._dynamical_matrix += dm_dd
 
     def _get_Gonze_dipole_dipole(self, q_red, q_direction):
+        num_atom = len(self._pcell)
         rec_lat = np.linalg.inv(self._pcell.cell)  # column vectors
         q_cart = np.array(np.dot(q_red, rec_lat.T), dtype="double")
         if q_direction is None:
@@ -725,9 +741,16 @@ class DynamicalMatrixGL(DynamicalMatrixNAC):
             )
             sys.exit(1)
 
+        if self._with_full_terms:
+            for i in range(num_atom):
+                C_recip[i, :, i, :] += self._dd_limiting
+            C_recip += self._get_real_dipole_dipole(q_red)
+            drift = self._dd_q0 + self._dd_limiting * num_atom + self._dd_real_q0
+            for i in range(num_atom):
+                C_recip[i, :, i, :] -= drift[i]
+
         # Mass weighted
         mass = self._pcell.masses
-        num_atom = len(self._pcell)
         for i in range(num_atom):
             for j in range(num_atom):
                 C_recip[i, :, j, :] *= 1.0 / np.sqrt(mass[i] * mass[j])
@@ -745,6 +768,10 @@ class DynamicalMatrixGL(DynamicalMatrixNAC):
         This is added to interpolated short range force constants
         to create full force constants. Called many times.
 
+        Returns
+        -------
+        shape=(num_atom, 3, num_atom, 3), dtype=complex
+
         """
         import phonopy._phonopy as phonoc
 
@@ -753,9 +780,14 @@ class DynamicalMatrixGL(DynamicalMatrixNAC):
         volume = self._pcell.volume
         dd = np.zeros((num_atom, 3, num_atom, 3), dtype=self._dtype_complex, order="C")
 
+        if self._with_full_terms:
+            dd_q0 = np.zeros((len(pos), 3, 3), dtype=self._dtype_complex, order="C")
+        else:
+            dd_q0 = self._dd_q0
+
         phonoc.recip_dipole_dipole(
             dd.view(dtype="double"),
-            self._dd_q0.view(dtype="double"),
+            dd_q0.view(dtype="double"),
             self._G_list,
             q_cart,
             q_dir_cart,
@@ -789,57 +821,47 @@ class DynamicalMatrixGL(DynamicalMatrixNAC):
             self._symprec,
         )
 
-        # Limiting contribution
-        # inv_eps = np.linalg.inv(self._dielectric)
-        # sqrt_det_eps = np.sqrt(np.linalg.det(self._dielectric))
-        # coef = (4.0 / 3 / np.sqrt(np.pi) * inv_eps
-        #         / sqrt_det_eps * self._Lambda ** 3)
-        # self._dd_q0 -= coef
-
-    def _get_py_dipole_dipole(self, K_list, q, q_dir_cart):
-        pos = self._pcell.positions
+    def _get_real_dipole_dipole(self, q_red):
         num_atom = len(self._pcell)
-        volume = self._pcell.volume
-        C = np.zeros((num_atom, 3, num_atom, 3), dtype=self._dtype_complex, order="C")
+        phase_all = np.exp(2j * np.pi * np.dot(self._svecs, q_red))
+        C_real = np.zeros((num_atom, 3, num_atom, 3), dtype=self._dtype_complex)
+        vals = (
+            -self._Lambda**3
+            * self._H
+            * phase_all
+            * np.linalg.det(self._dielectric) ** (-0.5)
+        )
+        for i_s in range(self._multi.shape[0]):
+            for i_p in range(self._multi.shape[1]):
+                m = self._multi[i_s, i_p]
+                C_real[self._s2pp_map[i_s], :, i_p, :] += (
+                    (vals[:, :, m[1]] + vals[:, :, m[1]].conj().T) / 2 / m[0]
+                )
+        return C_real
 
-        for q_K in K_list:
-            if np.linalg.norm(q_K) < self._symprec:
-                if q_dir_cart is None:
-                    continue
-                else:
-                    dq_K = q_dir_cart
-            else:
-                dq_K = q_K
+    def _run_real_dipole_dipole_q0(self):
+        self._dd_real_q0 = np.zeros((len(self._pcell), 3, 3), dtype=self._dtype_complex)
+        vals = -self._Lambda**3 * self._H * np.linalg.det(self._dielectric) ** (-0.5)
+        for i_s in range(self._multi.shape[0]):
+            for i_p in range(self._multi.shape[1]):
+                m = self._multi[i_s, i_p]
+                self._dd_real_q0[self._s2pp_map[i_s], :, :] += (
+                    (vals[:, :, m[1]] + vals[:, :, m[1]].conj().T) / 2 / m[0]
+                )
 
-            Z_mat = self._get_charge_sum(
-                num_atom, dq_K, self._born
-            ) * self._get_constant_factor(
-                dq_K, self._dielectric, volume, self._unit_conversion
-            )
-            for i in range(num_atom):
-                dpos = -pos + pos[i]
-                phase_factor = np.exp(2j * np.pi * np.dot(dpos, q_K))
-                for j in range(num_atom):
-                    C[i, :, j, :] += Z_mat[i, j] * phase_factor[j]
+    def _run_limiting_dipole_dipole(self):
+        """Calculate limiting contribution.
 
-        for q_K in K_list:
-            q_G = q_K - q
-            if np.linalg.norm(q_G) < self._symprec:
-                continue
-            Z_mat = self._get_charge_sum(
-                num_atom, q_G, self._born
-            ) * self._get_constant_factor(
-                q_G, self._dielectric, volume, self._unit_conversion
-            )
-            for i in range(num_atom):
-                C_i = np.zeros((3, 3), dtype=self._dtype_complex, order="C")
-                dpos = -pos + pos[i]
-                phase_factor = np.exp(2j * np.pi * np.dot(dpos, q_G))
-                for j in range(num_atom):
-                    C_i += Z_mat[i, j] * phase_factor[j]
-                C[i, :, i, :] -= C_i
+        Calculated only once.
 
-        return C
+        shape=(3, 3)
+
+        """
+        inv_eps = np.linalg.inv(self._dielectric)
+        sqrt_det_eps = np.sqrt(np.linalg.det(self._dielectric))
+        self._dd_limiting = (
+            -4.0 / 3 / np.sqrt(np.pi) * inv_eps / sqrt_det_eps * self._Lambda**3
+        )
 
     def _get_G_list(self, G_cutoff, g_rad=100):
         rec_lat = np.linalg.inv(self._pcell.cell)  # column vectors
@@ -858,10 +880,13 @@ class DynamicalMatrixGL(DynamicalMatrixNAC):
     def _get_H(self):
         lat = self._scell.cell
         cart_vecs = np.dot(self._svecs, lat)
-        Delta = np.dot(cart_vecs, np.linalg.inv(self._dielectric).T)
-        D = np.sqrt(cart_vecs * Delta).sum(axis=3)
+        eps_inv = np.linalg.inv(self._dielectric)
+        Delta = np.dot(cart_vecs, eps_inv.T)  # (N, 3)
+        D = np.sqrt((cart_vecs * Delta).sum(axis=1))  # (N,)
         x = self._Lambda * Delta
         y = self._Lambda * D
+        condition = y < 1e-10
+        y[condition] = 1  # dummy to avoid divergence
         y2 = y**2
         y3 = y**3
         exp_y2 = np.exp(-y2)
@@ -878,16 +903,15 @@ class DynamicalMatrixGL(DynamicalMatrixNAC):
             for i in np.ndindex(y.shape):
                 erfc_y[i] = erfc(y[i])
 
-        with np.errstate(divide="ignore", invalid="ignore"):
-            A = (3 * erfc_y / y3 + 2 / np.sqrt(np.pi) * exp_y2 * (3 / y2 + 2)) / y2
-            A[A == np.inf] = 0
-            A = np.nan_to_num(A)
-            B = erfc_y / y3 + 2 / np.sqrt(np.pi) * exp_y2 / y2
-            B[B == np.inf] = 0
-            B = np.nan_to_num(B)
+        A = np.where(
+            condition,
+            0,
+            (3 * erfc_y / y3 + 2 / np.sqrt(np.pi) * exp_y2 * (3 / y2 + 2)) / y2,
+        )
+        B = np.where(condition, 0, erfc_y / y3 + 2 / np.sqrt(np.pi) * exp_y2 / y2)
         H = np.zeros((3, 3) + y.shape, dtype="double", order="C")
         for i, j in np.ndindex((3, 3)):
-            H[i, j] = x[:, :, :, i] * x[:, :, :, j] * A - eps_inv[i, j] * B
+            H[i, j, :] = x[:, i] * x[:, j] * A - eps_inv[i, j] * B
         return H